[v0.87][tools] Split oversized pr_cmd.rs lifecycle module by responsibility (Closes #1367)
